+++ src/external/evented.ts
