+++ src/types/plugin.ts
@@ -1,5 +1,3 @@
-import type { MapPointerEvent } from '@2gis/mapgl/types';
-
 export type ColorModelString = `${'rgb' | 'hsl'}(${string})`;
 export type HexColorString = `#${string}`;
 /**
@@ -122,30 +120,4 @@ export interface ModelOptions {
      * List of buildings' identifiers that should be hidden
      */
     linkedIds?: string[];
-}
-
-/**
- * The list of events that can be emitted by the glTF plugin instance
- */
-export interface GltfPluginEventTable {
-    /**
-     * Emitted when model is clicked
-     */
-    clickModel: MapPointerEvent;
-    /**
-     * Emitted when poi is clicked
-     */
-    clickPoi: MapPointerEvent;
-    /**
-     * Emitted when the user moves the pointer over the poi.
-     */
-    mousemovePoi: MapPointerEvent;
-    /**
-     * Emitted when the user hovers over the poi.
-     */
-    mouseoverPoi: MapPointerEvent;
-    /**
-     * Emitted when the user moves the mouse away from the poi.
-     */
-    mouseoutPoi: MapPointerEvent;
 }
