+++ src/types/utils.ts
@@ -1,0 +1,1 @@
+export type Parameter<T extends (...args: any) => any> = Parameters<T>['0'];
