+++ src/utils/common.ts
@@ -1,0 +1,9 @@
+export function clamp(value: number, min: number, max: number): number {
+    value = Math.max(value, min);
+    value = Math.min(value, max);
+    return value;
+}
+
+export function degToRad(degrees: number): number {
+    return (degrees * Math.PI) / 180;
+}
